+++ conf/dataset/segmentation/s3dis/s3dis1x1.yaml
@@ -1,6 +1,6 @@
 # @package dataset
 defaults:
   - segmentation/default
-_target_: torch_points3d.dataset.s3dis1x1.s3dis_data_module
+_target_: torch_points3d.datasets.s3dis1x1.s3dis_data_module
 cfg:
   fold: 5
